Tests don't see my list. I want to kill unit testing
