+++ hw2/List/List.Tests/Properties/AssemblyInfo.cs
@@ -2,18 +2,18 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-[assembly: AssemblyTitle("ListTests")]
+[assembly: AssemblyTitle("List.Tests")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("ListTests")]
+[assembly: AssemblyProduct("List.Tests")]
 [assembly: AssemblyCopyright("Copyright ©  2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
 [assembly: ComVisible(false)]
 
-[assembly: Guid("d1c11b93-a907-4cbf-9258-e1be8f03b0c5")]
+[assembly: Guid("06ee3472-c6fb-4bea-ac55-de311c3cffa6")]
 
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("1.0.0.0")]
