--- hw2/List/ListTests/ListTests.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.168
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ListTests", "ListTests\ListTests.csproj", "{D1C11B93-A907-4CBF-9258-E1BE8F03B0C5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D1C11B93-A907-4CBF-9258-E1BE8F03B0C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D1C11B93-A907-4CBF-9258-E1BE8F03B0C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D1C11B93-A907-4CBF-9258-E1BE8F03B0C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D1C11B93-A907-4CBF-9258-E1BE8F03B0C5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D5DA6C85-12A3-4F08-86BE-A33B33C7826A}
-	EndGlobalSection
-EndGlobal
